buanderie-poller-tplink/requirements.txt
### Dependencies ###
pyHS100==0.3.0
google-cloud-datastore==1.3.0